feat: refactor into component and examples
